Semi-fix library being resized after sharing
For some reason, UIKit resizes the fully custom presented view controller. When dismissing a full-screen share action (e.g. Mail) the library is resized to full-screen.

There seem to be no callbacks or hook-ins for the presentation controller (e.g. in `viewWillTransition(to:)`)

A workaround is to reset the size in `containerViewWillLayoutSubviews` but that leaves a noticeable delay until the view is back to its intended size (`containerViewWillLayoutSubviews` is only called after the share dismiss animation completes).